+++ src/chart.cpp
@@ -22,7 +22,7 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
 
-#include "psschart.h"
+#include "chart.h"
 #include "stringutil.h"
 
 #include <QDateTime>
